+++ src/uQCme/__init__.py
@@ -1,5 +1,5 @@
 """uQCme - Microbial Quality Control Dashboard"""
 
-__version__ = "0.8.7"
+__version__ = "0.9.0"
 __description__ = "Microbial Quality Control Dashboard"
 __author__ = "SSI-DK"